Refactor name change quoted -> quote_table_name
Planning on using this in many more places in order to resolve a bug
where a camelCaseTableName breaks the csv interactive edit. In psql for
SELECT and other queries you need to double-quote wrap anything that has
any capital letters. This is because by default psql will transform any
query text to lowercase. Eg)
- SeLeCt * from POTAto; -> select * from potato;